fix canvas rendering after folder restructuring
investigate and fix rendering issues after reorganizing project structure
