+++ snake-react/src/App.css
@@ -3,7 +3,15 @@
   justify-content: center;
   margin-top: 1.5rem;
 }
+
 canvas {
   max-width: 90vw;
-  height: auto;
+  background-color: #fce4ec;
+}
+
+.title {
+  margin-bottom: 1.5rem;
+  text-align: center;
+  font-size: 2rem;
+  color: #ff0099;
 }
